pyproject.toml
[tool.poetry]
name = "flask_app"
version = "0.1.0"
description = ""
authors = ["Caleb Crouse"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.121.1"
constructs = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
